netnook/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.27
pydantic==2.6.1
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pytest==7.4.3
httpx==0.27.0
